requirement.txt
numpy==1.26.0
pandas==2.0.3
openpyxl==3.1.2
python-dotenv==1.0.1